Rename codex and library capabilities
