+++ src/lingtai/capabilities/__init__.py
@@ -14,11 +14,14 @@ if TYPE_CHECKING:
 # relative names.
 _BUILTIN: dict[str, str] = {
     # Always-on floor (lingtai.core)
-    "codex": "lingtai.core.codex",
+    "library": "lingtai.core.library",
+    "skills": "lingtai.core.skills",
+    # Deprecated compatibility names. Agent config normalization maps these
+    # before setup; direct callers still resolve for one release window.
+    "codex": "lingtai.core.library",
     "bash": "lingtai.core.bash",
     "avatar": "lingtai.core.avatar",
     "daemon": "lingtai.core.daemon",
-    "library": "lingtai.core.library",
     "mcp": "lingtai.core.mcp",
     "read": "lingtai.core.read",
     "write": "lingtai.core.write",
@@ -34,6 +37,75 @@ _BUILTIN: dict[str, str] = {
 _GROUPS: dict[str, list[str]] = {
     "file": ["read", "write", "edit", "glob", "grep"],
 }
+
+
+def canonical_capability_name(name: str) -> str:
+    """Return the canonical post-rename capability name for *name*.
+
+    Compatibility mapping:
+    - ``codex`` (old durable knowledge) -> ``library``
+    - ``library`` keeps its new meaning here; raw init/preset manifests that
+      predate the rename are normalized by ``normalize_capabilities`` below so
+      old ``library.paths`` becomes ``skills.paths`` before setup.
+    """
+    return "library" if name == "codex" else name
+
+
+def normalize_capabilities(capabilities: dict[str, dict]) -> dict[str, dict]:
+    """Normalize old capability names to the post-rename surface.
+
+    Old manifests commonly contain both ``codex`` and ``library``. In that
+    context the old ``library`` entry is the skill catalog, so it becomes
+    ``skills`` while old ``codex`` becomes the new knowledge ``library``.
+    New manifests should use ``library`` for knowledge and ``skills`` for the
+    skill catalog. If both old and new keys are present, canonical keys win and
+    missing ``paths`` from the old skill-catalog key are merged into ``skills``.
+    """
+    out: dict[str, dict] = {}
+    has_old_codex = "codex" in capabilities
+
+    def merge_dict(dst: str, value: object) -> None:
+        if value is None:
+            value = {}
+        if dst not in out:
+            out[dst] = value if isinstance(value, dict) else value  # type: ignore[assignment]
+            return
+        if isinstance(out[dst], dict) and isinstance(value, dict):
+            merged = dict(value)
+            merged.update(out[dst])
+            # Preserve skill path extras from either spelling.
+            if dst == "skills":
+                paths = []
+                seen = set()
+                for source in (value.get("paths", []), out[dst].get("paths", [])):
+                    if not isinstance(source, list):
+                        continue
+                    for p in source:
+                        if isinstance(p, str) and p not in seen:
+                            paths.append(p)
+                            seen.add(p)
+                if paths:
+                    merged["paths"] = paths
+            out[dst] = merged
+
+    for name, kwargs in capabilities.items():
+        if name == "codex":
+            target = "library"
+        elif (
+            name == "library"
+            and "skills" not in capabilities
+            and (
+                has_old_codex
+                or (isinstance(kwargs, dict) and "paths" in kwargs)
+            )
+        ):
+            # Old pair: codex + library, or old library-only config carrying
+            # path extras. The library entry was the skill catalog.
+            target = "skills"
+        else:
+            target = name
+        merge_dict(target, kwargs)
+    return out
 
 
 def expand_groups(names: list[str]) -> list[str]:
@@ -82,8 +154,9 @@ def get_all_providers() -> dict[str, dict]:
         "file": "lingtai.core.read",
         "bash": "lingtai.core.bash",
         "web_search": ".web_search",
-        "codex": "lingtai.core.codex",
         "library": "lingtai.core.library",
+        "skills": "lingtai.core.skills",
+        "codex": "lingtai.core.library",
         "vision": ".vision",
         "avatar": "lingtai.core.avatar",
         "daemon": "lingtai.core.daemon",
