+++ src/lingtai/core/__init__.py
@@ -1,8 +1,9 @@
-"""Always-on agent floor: file I/O, library, codex, daemon, avatar, bash.
+"""Always-on agent floor: file I/O, library, skills, daemon, avatar, bash.
 
-These ten capabilities form the baseline every functional agent uses. They are
+These capabilities form the baseline every functional agent uses. They are
 discovered through the registry in ``lingtai.capabilities.__init__`` (which points
 at this subpackage by absolute path), so dispatch and group expansion logic stays
-unchanged. This package exists to make the always-on tier visible in the import
-graph; it has no behavior of its own.
+unchanged. ``codex`` remains as a deprecated compatibility wrapper for the renamed
+knowledge ``library`` capability. This package exists to make the always-on tier
+visible in the import graph; it has no behavior of its own.
 """
