+++ src/lingtai/core/skills/manual/assets/skill-template.md
