+++ src/lingtai/i18n/en.json
@@ -26,14 +26,14 @@
   "bash.working_dir": "Working directory for the command (optional)",
   "bash.async": "Run command in background and return immediately with a job_id (default: false, only for action='run')",
   "bash.job_id": "Job ID for poll/cancel actions (returned by async run)",
-  "codex.description": "Your durable journal across molts — what you've learned, decided, and discovered. Each entry's id + title + summary is visible in your system prompt; content and supplementary load on demand via view. Use progressive disclosure: keep summaries concise and scannable (1-3 sentences); put detailed explanations in content; raw data, citations, and long-form reference in supplementary. This layered approach keeps your prompt efficient while preserving depth. Slots are limited — consolidate related entries when you fill up. submit to add. view to read. consolidate to merge. delete to remove.",
-  "codex.action": "submit: add a new entry (title + summary required; content + supplementary optional). view: read entry content by ID (requires ids; pass include_supplementary=true for backing material). consolidate: merge entries into one (requires ids, title, summary; content + supplementary optional). delete: remove entries (requires ids).",
-  "codex.title": "Entry title — one line. Required for submit and consolidate.",
-  "codex.summary": "Entry summary — 1-3 sentences. Always visible in your system prompt, so make it the part you'd want to see at a glance. Required for submit and consolidate.",
-  "codex.content": "Main body — durable prose, written when you submit. Optional. Up to ~500 words is a good shape; longer goes in supplementary. Read on demand via view.",
-  "codex.supplementary": "Backing material — citations, source data, the long form behind the principle. Optional for submit and consolidate. Read only when you call view with include_supplementary=true.",
-  "codex.ids": "Entry IDs — for view, consolidate, and delete.",
-  "codex.include_supplementary": "When true on view, also returns supplementary alongside content. Default: false.",
+  "library.description": "Your durable long-term knowledge library across molts — what you've learned, decided, and discovered. Each entry's id + title + summary is visible in your system prompt; content and supplementary material load on demand via view. Use progressive disclosure: keep summaries concise and scannable (1-3 sentences); put detailed explanations in content; raw data, citations, and long-form reference in supplementary. This layered approach keeps your prompt efficient while preserving depth. Slots are limited — consolidate related entries when you fill up. submit to add. view to read. consolidate to merge. delete to remove.",
+  "library.action": "submit: add a new entry (title + summary required; content + supplementary optional). view: read entry content by ID (requires ids; pass include_supplementary=true for backing material). consolidate: merge entries into one (requires ids, title, summary; content + supplementary optional). delete: remove entries (requires ids).",
+  "library.title": "Entry title — one line. Required for submit and consolidate.",
+  "library.summary": "Entry summary — 1-3 sentences. Always visible in your system prompt, so make it the part you'd want to see at a glance. Required for submit and consolidate.",
+  "library.content": "Main body — durable prose, written when you submit. Optional. Up to ~500 words is a good shape; longer goes in supplementary. Read on demand via view.",
+  "library.supplementary": "Backing material — citations, source data, the long form behind the principle. Optional for submit and consolidate. Read only when you call view with include_supplementary=true.",
+  "library.ids": "Entry IDs — for view, consolidate, and delete.",
+  "library.include_supplementary": "When true on view, also returns supplementary alongside content. Default: false.",
   "avatar.description": "Spawn a 他我 (alter ego) — a fully independent agent process. Inherits your init.json. For preset-equipped agents: boots on DEFAULT preset. Once spawned, it is detached — a new life, independent. Use mail or email to communicate. type='shallow' (default): blank slate, only init.json. type='deep': copies lingtai, pad, and codex. All spawns recorded in delegates/ledger.jsonl. IMPORTANT: the standard 'reasoning' parameter you write on this tool call automatically becomes the avatar's starting prompt — so write it as a thorough mission briefing, not just a rationale for the call. Actions: 'spawn': create new 他我. 'rules': set network-wide rules (requires karma). WARNING: each spawn creates an independent process consuming resources until sleep'd or suspended. Double-check reasoning before invoking. Use dry_run=true to preview.",
   "avatar.name": "True name (真名) for the 他我 (required, e.g. 'researcher', 'analyst', '学者'). This doubles as the avatar's working-directory basename under .lingtai/, so it must be a bare single segment: letters (any script — Latin, CJK, Cyrillic, etc.), digits, underscore, or hyphen only. No slashes, no dots, no spaces, no leading '.', max 64 chars. The avatar's display name (nickname) can be set separately and has no such constraints.",
   "avatar.type": "Spawn type. 'shallow' (default, 初生): blank slate — only init.json (LLM config + capabilities) inherited. 'deep' (二重身): full copy of character, pad, and codex.",
@@ -58,7 +58,7 @@
   "daemon.tasks.preset": "Optional preset file path. Must be a .json/.jsonc path as returned by system(action='presets'). Do NOT use shorthand names — use the full 'name' field from the presets listing. Example: '~/.lingtai-tui/presets/saved/cheap.json'. Omit to inherit the parent's tool surface.",
   "daemon.id": "Subagent ID for 'ask' action (e.g. 'em-1')",
   "daemon.message": "Follow-up message for 'ask' action",
-  "library.description": "Your per-agent skill library. The <available_skills> catalog in your system prompt lists every skill reachable right now. Call info to read the full library workflow (authoring, publishing, loading via pad.append) and verify your library is healthy. See also the library-manual skill in .library/intrinsic/.",
-  "library.action_info": "info: return the library-manual skill body plus a runtime health snapshot (catalog size, resolved paths, problems).",
-  "library.preamble": "The following skills provide specialized instructions for specific tasks. When a task matches a skill's description, use the read tool to load the SKILL.md file at the path shown in that skill's location field, then follow its instructions. Each skill folder may contain supporting files (scripts, templates, data) — the SKILL.md will reference them. Skills you use often can be pinned into your pad via psyche({object:'pad', action:'append', files:[location]}) for cross-turn persistence. If no skill clearly applies, do not load any."
+  "skills.description": "Your per-agent skill catalog. The <available_skills> catalog in your system prompt lists every skill reachable right now. Call info to read the full skills workflow (authoring, publishing, loading via pad.append) and verify your skills catalog is healthy. See also the skills-manual skill in .library/intrinsic/.",
+  "skills.action_info": "info: return the skills-manual skill body plus a runtime health snapshot (catalog size, resolved paths, problems).",
+  "skills.preamble": "The following skills provide specialized instructions for specific tasks. When a task matches a skill's description, use the read tool to load the SKILL.md file at the path shown in that skill's location field, then follow its instructions. Each skill folder may contain supporting files (scripts, templates, data) — the SKILL.md will reference them. Skills you use often can be pinned into your pad via psyche({object:'pad', action:'append', files:[location]}) for cross-turn persistence. If no skill clearly applies, do not load any."
 }
