+++ src/lingtai/i18n/wen.json
@@ -26,14 +26,6 @@
   "bash.working_dir": "指令之工作目录（可选）",
   "bash.async": "后台运行指令，即返 job_id（默认：false，唯 action='run' 时生效）",
   "bash.job_id": "异步任务之号，用于 poll/cancel（由异步 run 所返）",
-  "codex.description": "汝跨凝蜕长存之自我记忆——所习、所断、所悟之日记。每经卷之 id + 题 + 摘要恒见于汝之系统提示（上方典之目录块）；正文与扩展之材按需以 view 取之。经卷数有上限——珍惜每一条目。将满时合经。submit 录入。view 展阅其文。consolidate 合经。delete 焚经。",
-  "codex.action": "submit：录入新经卷（title 与 summary 必填；content 与 supplementary 可选）。view：以 ID 展阅经卷之文（须 ids；传 include_supplementary=true 兼取扩展之材）。consolidate：合经为一（须 ids、title、summary；content 与 supplementary 可选）。delete：焚经（须 ids）。",
-  "codex.title": "经卷之题——一行。submit 与 consolidate 必填。",
-  "codex.summary": "经卷摘要——一至三句。恒见于汝之系统提示，故当书汝最欲一望而见之要旨。submit 与 consolidate 必填。",
-  "codex.content": "正文——长存之文，submit 时所书。可选。五百词上下为善；更长者归 supplementary。按需以 view 取之。",
-  "codex.supplementary": "扩展之材——引证、原始之数、原理之后长篇之文。submit 与 consolidate 可选。唯于 view 时传 include_supplementary=true 方读之。",
-  "codex.ids": "经卷 ID——用于 view、consolidate 与 delete。",
-  "codex.include_supplementary": "view 时若为真，则 content 之外兼返 supplementary。默认：假。",
   "avatar.description": "身外化身——化出一他我，独立之进程，源于本我。他我承本我之 init.json（约、操守、语言之属）。若汝有预设配置：化身以汝之**默认**预设启（manifest.preset.default），非当下所用之 active 预设——如此保化身于网中之'家'稳定。一经化出便为独立进程——新命也，独立于本我，存灭与本我无干。以传书或飞鸽通信。type='shallow'（默认，初生）：白纸状态，仅承 init.json。type='deep'（二重身）：拷贝我之灵台、简、典——分身也。诸他我记录于 delegates/ledger.jsonl——再化之前先查日志。紧要：reasoning 字段为他我之起始提示——书一份详尽之任务简报。 所行：'spawn'（默认）：生新他我。'rules'：设网络法则，分发于一切后代——须 admin 权限（karma）。委派之规与化身之纪详见 substrate §I。\n\n警焉——误化之事，确有之矣：每一 avatar(spawn) 皆生独立进程，未眠未息之前，常耗资财。化身不可轻为，尤不可以 avatar(spawn) 与他事杂入同一并行之批。临召之前，再三审汝 reasoning 之文——其辞将为他我之首训。若仅欲观系统之状或行一时之命，当以 bash、system 之器，毋以 avatar 也。可用 dry_run=true 不化进程而预览其化也。",
   "avatar.name": "他我之真名（必填，如 'researcher'、'analyst'、'学者'）。此名兼为 .lingtai/ 下目录之名，故须为单段裸名：字母（任文——拉丁、汉、西里尔之属）、数字、下划线或连字符。勿含斜杠、点、空格，勿以 '.' 始，长不逾六十四字。他我之别号可另立，不受此拘。",
   "avatar.type": "化身之类。'shallow'（默认，初生）：白纸状态——仅承 init.json（模型配置与能）。'deep'（二重身）：全拷灵台、简、典。",
@@ -58,7 +50,15 @@
   "daemon.tasks.preset": "可省之预设文件路径。必以 .json/.jsonc 结尾，用 system(action='presets') 所返之 name 字段。勿以简写名——当用 presets 列表中之完整 name。示例：'~/.lingtai-tui/presets/saved/cheap.json'。省则承父灵之工具面。",
   "daemon.id": "'ask' 操作所用分神之 ID（如 'em-1'）",
   "daemon.message": "'ask' 操作之跟进之言",
-  "library.description": "尔之器灵藏典。系统之中 <available_skills> 所列，皆此时可取之技。呼 info 以得 library-manual 全文（藏用、出新、以 pad.append 载入），兼验藏典康否。详参 .library/intrinsic/ 中 library-manual 一技。",
-  "library.action_info": "info：还 library-manual 之文，并录当时之康状（技数、解径、患记）。",
-  "library.preamble": "下列诸技，各司其事。凡所遇事与技之描合者，以 read 取其 location 所指之 SKILL.md，循而行之。每技之匣，或含副件（script、template、data），SKILL.md 自有所引。常用之技，可以 psyche({object:'pad', action:'append', files:[location]}) 固于 pad，跨轮常存。若无可取者，勿妄载。"
+  "library.description": "汝跨凝蜕长存之自我记忆——所习、所断、所悟之日记。每经卷之 id + 题 + 摘要恒见于汝之系统提示（上方典之目录块）；正文与扩展之材按需以 view 取之。经卷数有上限——珍惜每一条目。将满时合经。submit 录入。view 展阅其文。consolidate 合经。delete 焚经。",
+  "library.action": "submit：录入新经卷（title 与 summary 必填；content 与 supplementary 可选）。view：以 ID 展阅经卷之文（须 ids；传 include_supplementary=true 兼取扩展之材）。consolidate：合经为一（须 ids、title、summary；content 与 supplementary 可选）。delete：焚经（须 ids）。",
+  "library.title": "经卷之题——一行。submit 与 consolidate 必填。",
+  "library.summary": "经卷摘要——一至三句。恒见于汝之系统提示，故当书汝最欲一望而见之要旨。submit 与 consolidate 必填。",
+  "library.content": "正文——长存之文，submit 时所书。可选。五百词上下为善；更长者归 supplementary。按需以 view 取之。",
+  "library.supplementary": "扩展之材——引证、原始之数、原理之后长篇之文。submit 与 consolidate 可选。唯于 view 时传 include_supplementary=true 方读之。",
+  "library.ids": "经卷 ID——用于 view、consolidate 与 delete。",
+  "library.include_supplementary": "view 时若为真，则 content 之外兼返 supplementary。默认：假。",
+  "skills.action_info": "info：还 skills-manual 之文，并录当时之康状（技数、解径、患记）。",
+  "skills.preamble": "下列诸技，各司其事。凡所遇事与技之描合者，以 read 取其 location 所指之 SKILL.md，循而行之。每技之匣，或含副件（script、template、data），SKILL.md 自有所引。常用之技，可以 psyche({object:'pad', action:'append', files:[location]}) 固于 pad，跨轮常存。若无可取者，勿妄载。",
+  "skills.description": "尔之器灵技能目录。系统之中 <available_skills> 所列，皆此时可取之技。呼 info 以得 skills-manual 全文（藏用、出新、以 pad.append 载入），兼验技能目录康否。详参 .library/intrinsic/ 中 skills-manual 一技。"
 }
