+++ src/lingtai/i18n/zh.json
@@ -26,14 +26,14 @@
   "bash.working_dir": "命令的工作目录（可选）",
   "bash.async": "后台运行命令并立即返回 job_id（默认：false，仅 action='run' 时生效）",
   "bash.job_id": "异步任务 ID，用于 poll/cancel（由异步 run 返回）",
-  "codex.description": "你的跨凝蜕长存自我之记忆——记录所学、所断、所悟之日记。每条之 id + 标题 + 摘要始终于系统提示中可见；正文与扩展材料按需以 view 取之。条目有上限——珍惜每一条。将满时合经。submit 录入。view 展阅。consolidate 合经。delete 焚经。",
-  "codex.action": "submit：添加新条目（title 与 summary 必填；content 与 supplementary 可选）。view：按 ID 读取条目内容（需要 ids；传 include_supplementary=true 同时获取扩展材料）。consolidate：合并条目为一（需要 ids、title、summary；content 与 supplementary 可选）。delete：删除条目（需要 ids）。",
-  "codex.title": "条目标题——一行。submit 和 consolidate 必填。",
-  "codex.summary": "条目摘要——1-3 句话。始终于你的系统提示中可见，故当书你最希望一眼看到之要旨。submit 和 consolidate 必填。",
-  "codex.content": "正文——长存之文，submit 时所书。可选。约 500 词为佳；更长者归入 supplementary。按需通过 view 读取。",
-  "codex.supplementary": "扩展材料——引证、原始数据、原理背后之长篇。submit 和 consolidate 可选。仅当 view 调用时传 include_supplementary=true 才读取。",
-  "codex.ids": "条目 ID——用于 view、consolidate 和 delete。",
-  "codex.include_supplementary": "view 时若为 true，则同时返回 supplementary 与 content。默认：false。",
+  "library.description": "你的跨凝蜕长存自我之记忆——记录所学、所断、所悟之日记。每条之 id + 标题 + 摘要始终于系统提示中可见；正文与扩展材料按需以 view 取之。条目有上限——珍惜每一条。将满时合经。submit 录入。view 展阅。consolidate 合经。delete 焚经。",
+  "library.action": "submit：添加新条目（title 与 summary 必填；content 与 supplementary 可选）。view：按 ID 读取条目内容（需要 ids；传 include_supplementary=true 同时获取扩展材料）。consolidate：合并条目为一（需要 ids、title、summary；content 与 supplementary 可选）。delete：删除条目（需要 ids）。",
+  "library.title": "条目标题——一行。submit 和 consolidate 必填。",
+  "library.summary": "条目摘要——1-3 句话。始终于你的系统提示中可见，故当书你最希望一眼看到之要旨。submit 和 consolidate 必填。",
+  "library.content": "正文——长存之文，submit 时所书。可选。约 500 词为佳；更长者归入 supplementary。按需通过 view 读取。",
+  "library.supplementary": "扩展材料——引证、原始数据、原理背后之长篇。submit 和 consolidate 可选。仅当 view 调用时传 include_supplementary=true 才读取。",
+  "library.ids": "条目 ID——用于 view、consolidate 和 delete。",
+  "library.include_supplementary": "view 时若为 true，则同时返回 supplementary 与 content。默认：false。",
   "avatar.description": "化身出一个他我——完全独立的智能体进程，源于自身。他我继承你的 init.json（约、操守、语言等）。若汝有预设配置：化身以汝之**默认**预设启动（manifest.preset.default），非当前所用之 active 预设——如此保化身于网中之'家'稳定如初。一经化出便为独立进程——新的生命，独立于你，存灭与你无干。使用 mail 或 email 通信。type='shallow'（默认，初生）：白纸状态，仅继承 init.json。type='deep'（二重身）：复制你的灵台、手记和典集——一个分身。所有记录在 delegates/ledger.jsonl 中——化出前先查日志。重要：reasoning 字段成为他我的起始提示——写一份详尽的任务简报。 动作：'spawn'（默认）：创建新他我。'rules'：设网络规则并分发给所有后代——需 admin 权限（karma）。委派之规与化身纪律详见 substrate §I。\n\n警告——误化之事确有发生：每次 avatar(spawn) 都创建一个独立进程，在被 sleep 或 suspend 之前持续消耗资源。切勿轻率化身，绝不可将 avatar(spawn) 与无关工具调用并入同一并行批次。调用前请反复核对你的 reasoning 字段——那段文字会成为他我的首条提示。若仅需查看系统状态或执行一次性命令，请用 bash 或 system 工具，而非 avatar。可用 dry_run=true 在不创建进程的情况下预览一次化身。",
   "avatar.name": "他我的真名（必填，例如 'researcher'、'analyst'、'学者'）。此名同时作为 .lingtai/ 下的目录名，故须为单段裸名：字母（任何文字——拉丁、中日韩、西里尔等）、数字、下划线或连字符。不得含斜杠、点、空格或以 '.' 开头，最长 64 字符。他我的显示别号可另设，无此限制。",
   "avatar.type": "化身类型。'shallow'（默认，初生）：白纸状态——仅继承 init.json（LLM 配置 + 能力）。'deep'（二重身）：完整复制性格、手记和典集。",
@@ -58,7 +58,7 @@
   "daemon.tasks.preset": "可选预设文件路径。必须是 .json/.jsonc 路径，使用 system(action='presets') 返回的 name 字段。不要用简写名——用 presets 列表里的完整 name。示例：'~/.lingtai-tui/presets/saved/cheap.json'。省略则继承父灵工具面。",
   "daemon.id": "'ask' 操作的分神 ID（如 'em-1'）",
   "daemon.message": "'ask' 操作的跟进消息",
-  "library.description": "你的器灵专属技能库。系统提示词中的 <available_skills> 目录列出了当前所有可用的技能。调用 info 可阅读完整的技能库工作流（编写、发布、通过 pad.append 加载），并验证技能库的健康状态。也可直接查阅 .library/intrinsic/ 中的 library-manual 技能。",
-  "library.action_info": "info：返回 library-manual 技能正文，以及运行时健康快照（目录规模、解析后的路径、问题列表）。",
-  "library.preamble": "以下技能为特定任务提供专门指导。当任务匹配某个技能的描述时，使用 read 工具读取 location 字段所示路径下的 SKILL.md 文件，然后遵循其中指示。每个技能文件夹可能包含支持文件（脚本、模板、数据）——SKILL.md 会引用它们。经常使用的技能可通过 psyche({object:'pad', action:'append', files:[location]}) 固定到 pad 中，使其跨轮次持久存在。若没有明显匹配的技能，则不加载任何技能。"
+  "skills.description": "你的器灵专属技能目录。系统提示词中的 <available_skills> 目录列出了当前所有可用的技能。调用 info 可阅读完整的技能库工作流（编写、发布、通过 pad.append 加载），并验证技能库的健康状态。也可直接查阅 .library/intrinsic/ 中的 skills-manual 技能。",
+  "skills.action_info": "info：返回 skills-manual 技能正文，以及运行时健康快照（目录规模、解析后的路径、问题列表）。",
+  "skills.preamble": "以下技能为特定任务提供专门指导。当任务匹配某个技能的描述时，使用 read 工具读取 location 字段所示路径下的 SKILL.md 文件，然后遵循其中指示。每个技能文件夹可能包含支持文件（脚本、模板、数据）——SKILL.md 会引用它们。经常使用的技能可通过 psyche({object:'pad', action:'append', files:[location]}) 固定到 pad 中，使其跨轮次持久存在。若没有明显匹配的技能，则不加载任何技能。"
 }
