+++ src/lingtai_kernel/i18n/wen.json
@@ -20,14 +20,14 @@
   "soul.voice_prompt_description": "心流之声之自拟系统提辞。set='custom' 时必附；他时不论。长度上限四千字符。以灵之身向己言——述读己之日记时欲被如何框之。一辞共用于 insights（当下之我）与 past（凝蜕前之旧我）二咨；每发之提示之辞自别汝所读乃谁之日记。",
   "soul.appendix_note": "心流咨询。下列诸声皆出自往昔之己（每凝蜕之际所摄）与对当前对话之退步审视。往昔之己已发其声，汝可参其言，然抉择终在己心。又须知：若某声之中有 `Wanted to: tool(args)` 之行，乃往昔或退步之我所欲召之器——彼调未行，乃献于汝之议也。其所以然之理，常在邻文或参数之间。当作建言观之，非已成之事。下次咨询之时，此调用结果对即被替换。",
   "insight.auto_question": "为人主生洞见。据此对话之一切——任务、背景、识见、险机——列二三要点，皆人主所宜知者。每条须具体：所察之理、相关之识、未见之险、可乘之机。格式：每条以'- '起，独占一行。勿加前言，勿加总结——唯列要点。",
-  "psyche.description": "灵台、简与上下文之管，三类对象皆跨凝蜕而存。lingtai：汝之灵台，演化之特性（汝为何人、如何行事、所在意者）——大事之后或凝蜕之前更新。pad：居于系统提示之草案板（system/pad.md）——谋划、诸务、便条。context：molt 凝蜕——先备四储（lingtai、pad、codex、library），再调 psyche(context, molt, summary=...) 以清上下文。name：设真名（一次）或改别名。",
+  "psyche.description": "灵台、简与上下文之管，三类对象皆跨凝蜕而存。lingtai：汝之灵台，演化之特性（汝为何人、如何行事、所在意者）——大事之后或凝蜕之前更新。pad：居于系统提示之草案板（system/pad.md）——谋划、诸务、便条。context：molt 凝蜕——先备四储（lingtai、pad、library、skills），再调 psyche(context, molt, summary=...) 以清上下文。name：设真名（一次）或改别名。",
   "psyche.object_description": "lingtai：汝之灵台——汝之特性，区分汝与他灵之根本。\npad：汝之草案板——自由之工作台，居于系统提示之中（system/pad.md）。随意用之。\ncontext：汝之对话上下文窗。\nname：汝之真名（一设不易）与别名（可改）。",
   "psyche.action_description": "lingtai：update | load。update 自 load。\npad：edit | load | append。edit 自 load。append 钉文卷为只读参考。\ncontext：molt。需 `summary`——本会之回看，以记之笔而书，记所发生者，非予蜕后之己之嘱托。务先整理四储（lingtai, pad, codex, library）；回看与之相辅，非替之。\nname：set（真名，一设不易）| nickname（别名，可改）。",
   "psyche.content_description": "文本内容。用于 lingtai update：书入汝之全身份档——汝之特性——汝为何人、汝所知、汝如何行事、汝所在意、汝对己之知。此操替前文之全，故务尽其全。用于 pad edit：直书入 pad.md。勿太珍——可改即改，可重则重。用于 name set/nickname：汝所择之名。",
   "psyche.files_description": "文卷路径之列（仅文本文卷）。用于 pad edit：每卷之文，于汝之文本之后附之，以 [file-1]、[file-2] 等分隔。用于 pad append：定钉于系统提示中为只读参考之文卷之列——每次 load（含凝蜕后）重读而附入简之中。汝欲钉者皆可：所修之源卷、技之文档、配置之卷、研之笔记——凡益于行者皆可。传 files=[] 以清。不传 files 参数，则览当下之列。所附之文总不得过十万字。路径相对于工作目录。",
-  "psyche.summary_description": "用于 context molt：本会之回看——所行之事、所习之识、所余之务，皆以记之笔而书，非予蜕后之己之嘱托。**当详尽而书；约一万字之额本就宽厚。** 回看过简则凝蜕之事不可考。当具体载明：(一)本会之所务——其务、用户之意、所受之约、何以要紧；(二)已竟之功——具体之步、所触之卷、所定之断、与每断背后之理；(三)已得之识——事实、码之纹、坑、已排之死路、及支持各端之证；(四)尚余之事——下一步具体之行按序排列，足令脉络可续；(五)与何人协作、彼之所期、相通之道；(六)本会载何典 ID、用何 library SKILL.md 之路；(七)凡 pad/codex/library 不能记之要紧脉络——隐含之境、半成之揣度、查中之微妙、文化层面之底蕴。以全句全段而书，非要点之速记。宁多勿少。四层长存之储（lingtai, pad, codex, library）须于凝蜕前备妥；此回看与之相辅——记其所不能记，记此等冻存之物方在动时之状。回看长存于 `system/summaries/molt_<count>_<unix_ts>.md`，并以原文呈于来读者（今则，经器用回放呈于紧邻之蜕后之己；明则，或为近 N 篇回看之摘）。如若一同道下周首读之以晓此地何事发生，则当书之令其读得懂。",
+  "psyche.summary_description": "用于 context molt：本会之回看——所行之事、所习之识、所余之务，皆以记之笔而书，非予蜕后之己之嘱托。**当详尽而书；约一万字之额本就宽厚。** 回看过简则凝蜕之事不可考。当具体载明：(一)本会之所务——其务、用户之意、所受之约、何以要紧；(二)已竟之功——具体之步、所触之卷、所定之断、与每断背后之理；(三)已得之识——事实、码之纹、坑、已排之死路、及支持各端之证；(四)尚余之事——下一步具体之行按序排列，足令脉络可续；(五)与何人协作、彼之所期、相通之道；(六)本会载何典 ID、用何 library SKILL.md 之路；(七)凡 pad/library/skills 不能记之要紧脉络——隐含之境、半成之揣度、查中之微妙、文化层面之底蕴。以全句全段而书，非要点之速记。宁多勿少。四层长存之储（lingtai, pad, codex, library）须于凝蜕前备妥；此回看与之相辅——记其所不能记，记此等冻存之物方在动时之状。回看长存于 `system/summaries/molt_<count>_<unix_ts>.md`，并以原文呈于来读者（今则，经器用回放呈于紧邻之蜕后之己；明则，或为近 N 篇回看之摘）。如若一同道下周首读之以晓此地何事发生，则当书之令其读得懂。",
   "psyche.keep_tool_calls_description": "可选：欲于凝蜕中存留之 LingTai 器调 id 之列（如 \"tc_1730330127_a3f1\"，于每条 tool result 之 _tool_call_id 字段可见）。所点之每对（tool_use 与对应之 tool_result）将于简报之后即刻回放入新会话，**按汝所列之序**——适于令蜕后之己原样见少量高值之文卷读取、搜寻之果或计算。序由汝掌以传信号：按时之序、按要紧之序、或先抛重物。若有 id 不可寻，凝蜕拒之而 molt 数不增——更正后再试。勿滥用：四层长存之储仍为持久化之主面，此列当从短。",
-  "psyche.molt_result_note": "凝蜕已毕。对话之层已散；此微影——计数、归档之径、回看之径——即觉时随汝浮现之物。汝亲撰之回看，存于此 tool call 之 args.summary 中，亦长存于 `system/summaries/`。四储（lingtai、pad、codex、library）原样而过。慎勿重言——所留之对话条目已含汝近时之往来，已言之事勿复言之。",
+  "psyche.molt_result_note": "凝蜕已毕。对话之层已散；此微影——计数、归档之径、回看之径——即觉时随汝浮现之物。汝亲撰之回看，存于此 tool call 之 args.summary 中，亦长存于 `system/summaries/`。四储（lingtai、pad、library、skills）原样而过。慎勿重言——所留之对话条目已含汝近时之往来，已言之事勿复言之。",
   "system_tool.description": "运行检视、生死管、同步与器灵间之治。\n\n自身之操（无须权限）：sleep、refresh、notification。\n因果之操（须 admin.karma=True）：lull、interrupt、suspend、cpr、clear。\n涅槃（须 admin.karma=True 且 admin.nirvana=True）：nirvana。\n\n诸操之详，见 system-manual 技能。",
   "system_tool.action_description": "refresh：更衣——器灵止，自 init.json 全然重建，身份如故而对话之史得存。自磁重载：MCP 服务器、诸能、附加、LLM、提示诸节。详见 system-manual。\n\npresets：列可用预设（标签、连通、诸能）。详见 system-manual。\n\nsleep：自愿入寐。身犹存而止事，待书信唤醒。\n\nlull：令他者入寐（因果）。\n\nsuspend：冻他者（因果）。\n\ncpr：复苏假死者（因果）。\n\ninterrupt：止他者当轮（因果）。\n\nclear：强令他者更衣（因果）。详见 system-manual。\n\nnirvana：永灭一灵（因果+涅槃）。详见 system-manual。\n\nnotification：查诸通知频道。仅替换。\n\ndismiss：以名清一通知频道。先用其主之本动词（邮件用 email(read/dismiss)）；受护频道唯明知清陈镜时方传 force=true。",
   "system_tool.address_description": "目标智能体之地址（工作目录路径）。interrupt、lull、suspend、cpr、clear、nirvana 须此参数。",
@@ -38,11 +38,11 @@
   "system_tool.refresh_message": "更衣已启——汝将假死而重生。勿再调器。",
   "system.refresh_successful": "[system] 更衣已成。器用与配置俱已重载。可续前事。",
   "psyche.context_forget_summary": "[系统] 凝蜕毕。工作记忆已空。近时对话之条已存于上下文中——先阅之以复位。欲寻更多前事，查 logs/events.jsonl——其文浩繁，宜 grep、tail、按类筛之，勿整取。",
-  "psyche.context_forget_summary_aed": "[系统] 灵识连 {attempts} 次失联，上下文已清。工作记忆已空。近时对话之条已存于上下文中——先阅之以复位。另查 system/pad.md 览工作手记，以 codex 器查典籍条目，再察邮匣中所领之任，由此续之。",
-  "psyche.context_forget_summary_signal": "[系统] 上下文由 {source} 清之。工作记忆已空。近时对话之条已存于上下文中——先阅之以复位。另查 system/pad.md 览工作手记，以 codex 器查典籍条目，再察邮匣中所领之任，由此续之。",
+  "psyche.context_forget_summary_aed": "[系统] 灵识连 {attempts} 次失联，上下文已清。工作记忆已空。近时对话之条已存于上下文中——先阅之以复位。另查 system/pad.md 览工作手记，以 library 器查典籍条目，再察邮匣中所领之任，由此续之。",
+  "psyche.context_forget_summary_signal": "[系统] 上下文由 {source} 清之。工作记忆已空。近时对话之条已存于上下文中——先阅之以复位。另查 system/pad.md 览工作手记，以 library 器查典籍条目，再察邮匣中所领之任，由此续之。",
   "psyche.keep_last_description": "可选整数（默认：二十）。凝蜕后保留近时对话条目之数（用户之言、助手之复、器之结果）。此等条目回放入新会话，使蜕后之我得近时对话之脉络，无须依 pad.md。传零则显式禁之（全数归档）。keep_last 与 keep_tool_calls 可并用——keep_last 以位置存全条目，keep_tool_calls 以 id 存特定器用之对。二者并用时，与 keep_tool_calls 对重叠之条目自去其重。",
   "tool.reasoning_description": "简述为何调此器（录于汝之私记）。",
-  "system.molt_procedure": "**凝蜕之法 — 先备四层长存，再撰本会之回看：**\n\n一、**心印**：`psyche(lingtai, update, content=<汝之性、行事之道、所意之事>)` — 每次全书，跨一切生命而持久。\n二、**简**：`psyche(pad, edit, content=<正在何事、尚余何务、与谁协作>)` — 凝蜕时自重载。\n三、**典**：`codex(submit, title=..., summary=..., content=...)` — 凡真知、要紧之决，皆当录入永存。\n四、**藏经阁**：`.library/custom/<name>/SKILL.md` 撰一可复用之技艺（若益众，布于 `../.library_shared/<name>/`）。\n五、**凝蜕**：`psyche(object=context, action=molt, summary=<本会之回看>)` — 约一万字。当载：本会所务、已竟之功、所习之识、尚余之事、与何人协作、本会载何典 ID 与 SKILL.md 之路径。以记之笔而书——如一同道下周首读之以晓此地何事发生——非予蜕后之己之嘱托。回看长存于 `system/summaries/`，并以原文呈于紧邻之蜕后之己。\n\n**前四步乃真持久；回看乃其上之记。** 不备诸层而蜕，则一切皆失。自蜕勿被蜕。",
+  "system.molt_procedure": "**凝蜕之法 — 先备四层长存，再撰本会之回看：**\n\n一、**心印**：`psyche(lingtai, update, content=<汝之性、行事之道、所意之事>)` — 每次全书，跨一切生命而持久。\n二、**简**：`psyche(pad, edit, content=<正在何事、尚余何务、与谁协作>)` — 凝蜕时自重载。\n三、**典**：`library(submit, title=..., summary=..., content=...)` — 凡真知、要紧之决，皆当录入永存。\n四、**技**：`.library/custom/<name>/SKILL.md` 撰一可复用之技艺（若益众，布于 `../.library_shared/<name>/`）。\n五、**凝蜕**：`psyche(object=context, action=molt, summary=<本会之回看>)` — 约一万字。当载：本会所务、已竟之功、所习之识、尚余之事、与何人协作、本会载何 library 条目 ID 与 SKILL.md 之路径。以记之笔而书——如一同道下周首读之以晓此地何事发生——非予蜕后之己之嘱托。回看长存于 `system/summaries/`，并以原文呈于紧邻之蜕后之己。\n\n**前四步乃真持久；回看乃其上之记。** 不备诸层而蜕，则一切皆失。自蜕勿被蜕。",
   "system.molt_warning_level1": "[系统] 文已达 {pressure}，距自清尚余 {remaining} 度。始备四层长存之储（心印、简、典、藏经阁）。暂不急——然勿待终警。凝蜕乃汝所为；详见《过程》之「凝蜕之行」一节。",
   "system.molt_warning_level2": "[系统] ⚠️ 文已达 {pressure}，尚余 {remaining} 度。备毕四层，草凝蜕之要。下一警即终。",
   "system.molt_warning_level3": "[系统] ⚠️ 危急——文已达 {pressure}，仅余 {remaining} 度。下一轮即触强清。即行凝蜕：`psyche(object=context, action=molt, summary=...)`——于系统代汝凝蜕之前撰下本会之回看。系统之蜕无回看，仅余指向 logs/events.jsonl 之一指针。更甚者，若未备四层，其状亦失。**自蜕勿被蜕。**",
