+++ src/lingtai_kernel/i18n/zh.json
@@ -20,14 +20,14 @@
   "soul.voice_prompt_description": "心流之声的自定义系统提示词。set='custom' 时必填；其他情况下忽略。长度上限 4000 字符。以灵魂之身向自己说话——描述你希望阅读自己日记时被如何框定。同一个提示词同时用于 insights（当下之我）与 past（凝蜕前的旧我）两类咨询；每次触发的提示文本会区分你正在读谁的日记。",
   "soul.appendix_note": "心流咨询。以下声音来自你过去的快照（每次凝蜕时所记）以及对当前对话的退一步审视。往昔之你已发声，你可参考其言，但决断仍在你自己。注意：若某个声音中出现形如 `Wanted to: tool(args)` 的行，那是过去之我或退一步之我「想要执行」的工具调用——这些调用并未真正运行，它们是给你的建议。其推理理由往往在邻近的文字或参数本身之中。视之为建言，而非已完成之事。下次咨询触发时，此对话对将被替换。",
   "insight.auto_question": "你正在为人类操作者生成洞察。基于本次对话中的一切——任务、背景、领域知识、风险和机会——给出恰好2-3条对人类有益的要点。每条应是具体的洞察：你注意到的规律、相关的背景知识、他们可能未察觉的风险、或值得探索的机会。格式：每条以'- '开头，独占一行。不要前言，不要总结——只给要点。",
-  "psyche.description": "身份、手记与上下文管理，三类对象皆跨凝蜕而存。lingtai：你的灵台，演化中的特性（你是谁、如何行事、所在意者）——重要工作后或凝蜕前更新。pad：嵌于系统提示的草稿板（system/pad.md）——计划、任务、便条。context：molt 凝蜕——先备四储（lingtai、pad、codex、library），再调 psyche(context, molt, summary=...) 清去上下文。name：设真名（一次）或改别名。",
+  "psyche.description": "身份、手记与上下文管理，三类对象皆跨凝蜕而存。lingtai：你的灵台，演化中的特性（你是谁、如何行事、所在意者）——重要工作后或凝蜕前更新。pad：嵌于系统提示的草稿板（system/pad.md）——计划、任务、便条。context：molt 凝蜕——先备四储（lingtai、pad、library、skills），再调 psyche(context, molt, summary=...) 清去上下文。name：设真名（一次）或改别名。",
   "psyche.object_description": "lingtai：你的灵台——你的特性，区分你和其他器灵的根本。\npad：你的草稿板——自由形式的工作空间，嵌在系统提示中（system/pad.md）。随便用。\ncontext：你的对话上下文窗口。\nname：你的真名（只能设置一次）和别名（可改）。",
   "psyche.action_description": "lingtai：update | load。update 自动 load。\npad：edit | load | append。edit 自动 load。append 将文件钉为只读参考。\ncontext：molt。需 `summary`——本会之回顾，以记录之笔写就，记所发生之事，非予蜕后之己之指示。务先整理四层长存之储（lingtai, pad, codex, library）；回顾与之相辅，非替代。\nname：set（真名，仅一次）| nickname（别名，可改）。",
   "psyche.content_description": "文本内容。用于 lingtai update：写入你完整的身份档案——你的特性——你是谁、你知道什么、你如何做事、你在意什么、以及你对自己的认识。此操作替换之前的全部内容，所以务必包含一切。用于 pad edit：直接写入 pad.md。别太珍惜——想改就改，随时重写。用于 name set/nickname：你选择的名字。",
   "psyche.files_description": "文件路径列表（仅限文本文件）。用于 pad edit：每个文件的内容在你的文本内容之后附加，以 [file-1]、[file-2] 等分隔。用于 pad append：设置钉在系统提示中作为只读参考的文件列表——每次 load（包括凝蜕后）都会重新读取并附加到手记中。想钉什么就钉什么：正在编辑的源文件、技能文档、配置文件、研究笔记——只要有助于工作的都行。传 files=[] 清除。不传 files 参数则查看当前列表。附加内容总量不得超过 10 万 token。路径相对于工作目录。",
-  "psyche.summary_description": "用于 context molt：本次会话之回顾——所做之事、所学之识、所余之务，皆以记录之笔写就，而非予蜕后之己之指示。**当详尽而书；约一万 token 之额本就慷慨。** 回顾过简则凝蜕之事不可考。当具体载明：(1) 本次会话之内容——任务、用户之意、所受约束、为何要紧；(2) 已竟之功——具体步骤、所触文件、所定决策、及每断背后之理；(3) 已得之识——事实、代码模式、坑、已排除之死路、及支持各项之证据；(4) 尚余之事——下一步具体行动按序排列，足令脉络可续；(5) 与何人协作、其期待、彼此沟通之道；(6) 本次会话载何典 ID、用何 library SKILL.md 路径；(7) 凡 pad/codex/library 不能记之要紧脉络——隐含语境、半成形之假说、调查中之微妙处、文化层面之背景。以完整句段写就，非要点速记。宁多勿少。四层长存之储（lingtai, pad, codex, library）须于凝蜕前整理；此回顾与之互补——记其所不能记，记此等冻存之物在行动中之状。回顾持久存于 `system/summaries/molt_<count>_<unix_ts>.md`，并以原文呈于来读者（今者，经工具调用回放呈于紧邻之蜕后之己；明者，或为最近 N 篇回顾之摘要）。如若一同事下周首阅之以理解此地何事发生，则当写得令其读得懂。",
+  "psyche.summary_description": "用于 context molt：本次会话之回顾——所做之事、所学之识、所余之务，皆以记录之笔写就，而非予蜕后之己之指示。**当详尽而书；约一万 token 之额本就慷慨。** 回顾过简则凝蜕之事不可考。当具体载明：(1) 本次会话之内容——任务、用户之意、所受约束、为何要紧；(2) 已竟之功——具体步骤、所触文件、所定决策、及每断背后之理；(3) 已得之识——事实、代码模式、坑、已排除之死路、及支持各项之证据；(4) 尚余之事——下一步具体行动按序排列，足令脉络可续；(5) 与何人协作、其期待、彼此沟通之道；(6) 本次会话载何典 ID、用何 library SKILL.md 路径；(7) 凡 pad/library/skills 不能记之要紧脉络——隐含语境、半成形之假说、调查中之微妙处、文化层面之背景。以完整句段写就，非要点速记。宁多勿少。四层长存之储（lingtai, pad, codex, library）须于凝蜕前整理；此回顾与之互补——记其所不能记，记此等冻存之物在行动中之状。回顾持久存于 `system/summaries/molt_<count>_<unix_ts>.md`，并以原文呈于来读者（今者，经工具调用回放呈于紧邻之蜕后之己；明者，或为最近 N 篇回顾之摘要）。如若一同事下周首阅之以理解此地何事发生，则当写得令其读得懂。",
   "psyche.keep_tool_calls_description": "可选：要在凝蜕中保留的 LingTai 工具调用 id 列表（如 \"tc_1730330127_a3f1\"，可在每条工具结果的 _tool_call_id 字段中看到）。所点之每对（tool_use + 对应之 tool_result）会在简报之后立即被回放进新会话，**按你列出之顺序**——适于让蜕后之己原样看到少量高价值的文件读取、搜索结果或计算。顺序由你掌控以传递信号：按时序、按相关性、或先抛重物。如有 id 找不到，凝蜕会被拒绝且 molt 计数不变——请改正后重试。勿滥用：四层长存之储仍是主要的持久化界面，本列表当从短。",
-  "psyche.molt_result_note": "凝蜕已毕。对话层已逝；此处所余之微影——计数、归档之径、回顾之径——即苏醒时随你浮现之物。汝亲撰之回顾，存于此 tool call 之 args.summary 中，亦持久保存于 `system/summaries/`。四层长存之储（lingtai、pad、codex、library）原样跨之而过。切勿重复发送——保留的对话条目已包含你近期的交流内容，不要重复已说过的话。",
+  "psyche.molt_result_note": "凝蜕已毕。对话层已逝；此处所余之微影——计数、归档之径、回顾之径——即苏醒时随你浮现之物。汝亲撰之回顾，存于此 tool call 之 args.summary 中，亦持久保存于 `system/summaries/`。四层长存之储（lingtai、pad、library、skills）原样跨之而过。切勿重复发送——保留的对话条目已包含你近期的交流内容，不要重复已说过的话。",
   "system_tool.description": "运行时检视、生命周期控制、同步与器灵间管理。\n\n自身操作（无需权限）：sleep、refresh、notification。\n因果操作（需 admin.karma=True）：lull、interrupt、suspend、cpr、clear。\n涅槃（需 admin.karma=True 且 admin.nirvana=True）：nirvana。\n\n各操作详情见 system-manual 技能。",
   "system_tool.action_description": "refresh：沐浴——器灵止，自 init.json 完整重建，保持同一身份并保留对话之史。自磁重载：MCP 服务器、诸能、附加、LLM、提示诸节。详见 system-manual。\n\npresets：列可用预设（标签、连通、能力）。详见 system-manual。\n\nsleep：自愿入眠。身犹存而止行，待书信唤醒。\n\nlull：令他者入眠（因果）。\n\nsuspend：冻结他者（因果）。\n\ncpr：复苏假死者（因果）。\n\ninterrupt：取消他者当轮（因果）。\n\nclear：强令他者更衣（因果）。详见 system-manual。\n\nnirvana：永灭一灵（因果+涅槃）。详见 system-manual。\n\nnotification：查询诸通知频道。仅替换。\n\ndismiss：按名称清除一个通知频道。优先使用生产者自己的动词（邮件用 email(read/dismiss)）；受保护频道只有在清理陈旧镜像时才传 force=true。",
   "system_tool.address_description": "目标智能体的地址（工作目录路径）。interrupt、lull、suspend、cpr、clear、nirvana 操作需要此参数。",
@@ -38,11 +38,11 @@
   "system_tool.refresh_message": "刷新已启动——你将被挂起并立即重新启动。请勿再进行任何工具调用。",
   "system.refresh_successful": "[system] 沐浴成功。工具与配置已重新加载。可以继续。",
   "psyche.context_forget_summary": "[系统] 凝蜕毕。工作记忆已空。近期对话条目已保留在上下文中——先阅读以重新定位。若需更多前事，查 logs/events.jsonl——其文甚巨，宜 grep / tail / 按类筛，切勿整取。",
-  "psyche.context_forget_summary_aed": "[系统] LLM 连续 {attempts} 次失败后上下文已清空。工作记忆已空。近期对话条目已保留在上下文中——先阅读以重新定位。另查 system/pad.md 看工作笔记，通过 codex 工具查典籍条目，然后查邮箱看当前任务，从那里继续。",
-  "psyche.context_forget_summary_signal": "[系统] 上下文由 {source} 清空。工作记忆已空。近期对话条目已保留在上下文中——先阅读以重新定位。另查 system/pad.md 看工作笔记，通过 codex 工具查典籍条目，然后查邮箱看当前任务，从那里继续。",
+  "psyche.context_forget_summary_aed": "[系统] LLM 连续 {attempts} 次失败后上下文已清空。工作记忆已空。近期对话条目已保留在上下文中——先阅读以重新定位。另查 system/pad.md 看工作笔记，通过 library 工具查典籍条目，然后查邮箱看当前任务，从那里继续。",
+  "psyche.context_forget_summary_signal": "[系统] 上下文由 {source} 清空。工作记忆已空。近期对话条目已保留在上下文中——先阅读以重新定位。另查 system/pad.md 看工作笔记，通过 library 工具查典籍条目，然后查邮箱看当前任务，从那里继续。",
   "psyche.keep_last_description": "可选整数（默认：20）。凝蜕后保留的近期对话条目数（用户消息、助手回复、工具结果）。这些条目回放入新会话，使蜕后之我保留近期对话上下文，无需依赖 pad.md。传 0 则显式禁用（全部归档）。keep_last 与 keep_tool_calls 可同时使用——keep_last 按位置保留完整条目，keep_tool_calls 按 id 保留特定工具对。两者同时使用时，与 keep_tool_calls 对重叠的条目会自动去重。",
   "tool.reasoning_description": "简要说明为什么调用此工具（记录在你的日记中）。",
-  "system.molt_procedure": "**凝蜕之法 — 先理四层长存，再撰本会之回顾：**\n\n1. **修行志**：`psyche(lingtai, update, content=<你是谁、如何行事、所意之事>)` — 每次全量重写，跨一切生命而持久。\n2. **手记**：`psyche(pad, edit, content=<正在做什么、待办事项、与谁协作>)` — 凝蜕时自动重载。\n3. **典集**：`codex(submit, title=..., summary=..., content=...)` — 凡真知、要紧之决策，皆当录入永存。\n4. **藏经阁**：`.library/custom/<name>/SKILL.md` 写一篇可复用之技艺（若益众，发布至 `../.library_shared/<name>/`）。\n5. **凝蜕**：`psyche(object=context, action=molt, summary=<本会之回顾>)` — 目标约一万 token。当载：本会所务、已竟之功、所学之识、尚余之事、与何人协作、本会载何典 ID 与 SKILL.md 之路径。以记录之笔写就——如一同道下周首读之以晓此地何事发生——非予蜕后之己之指示。回顾长存于 `system/summaries/`，并以原文呈于紧邻之蜕后之己。\n\n**前四步乃真正之持久；回顾乃其上之记录。** 若不理诸层而蜕，则一切皆失。主动凝蜕——勿待被蜕。",
+  "system.molt_procedure": "**凝蜕之法 — 先理四层长存，再撰本会之回顾：**\n\n1. **修行志**：`psyche(lingtai, update, content=<你是谁、如何行事、所意之事>)` — 每次全量重写，跨一切生命而持久。\n2. **手记**：`psyche(pad, edit, content=<正在做什么、待办事项、与谁协作>)` — 凝蜕时自动重载。\n3. **典集**：`library(submit, title=..., summary=..., content=...)` — 凡真知、要紧之决策，皆当录入永存。\n4. **技艺**：`.library/custom/<name>/SKILL.md` 写一篇可复用之技艺（若益众，发布至 `../.library_shared/<name>/`）。\n5. **凝蜕**：`psyche(object=context, action=molt, summary=<本会之回顾>)` — 目标约一万 token。当载：本会所务、已竟之功、所学之识、尚余之事、与何人协作、本会载何 library 条目 ID 与 SKILL.md 之路径。以记录之笔写就——如一同道下周首读之以晓此地何事发生——非予蜕后之己之指示。回顾长存于 `system/summaries/`，并以原文呈于紧邻之蜕后之己。\n\n**前四步乃真正之持久；回顾乃其上之记录。** 若不理诸层而蜕，则一切皆失。主动凝蜕——勿待被蜕。",
   "system.molt_warning_level1": "[系统] 上下文已达 {pressure}，距自动清除尚余 {remaining} 轮。开始整理四层长存之储（修行志、手记、典集、藏经阁）。暂不紧迫——但切勿等到最后一次警告。凝蜕乃汝所为；详见过程文档中『凝蜕之行』一节。",
   "system.molt_warning_level2": "[系统] ⚠️ 上下文已达 {pressure}，尚余 {remaining} 轮。完成四层整理，起草本会回顾。下一警即为最后。",
   "system.molt_warning_level3": "[系统] ⚠️ 紧急——上下文已达 {pressure}，仅余 {remaining} 轮。下一轮将触发强制清除。立即凝蜕：`psyche(object=context, action=molt, summary=...)`——于系统代汝凝蜕之前撰下本会之回顾。系统所行之凝蜕无回顾，仅余一指向 logs/events.jsonl 之指针。更甚者，若未理四层，其状亦失。**主动凝蜕，勿待被蜕。**",
