+++ src/lingtai_kernel/presets.py
@@ -36,7 +36,6 @@ the kernel surfaces the whole object verbatim to the agent.
 from __future__ import annotations
 
 import logging
-import os
 from pathlib import Path
 
 log = logging.getLogger(__name__)
@@ -295,10 +294,11 @@ def materialize_active_preset(data: dict, working_dir: Path) -> None:
     so downstream validators and consumers see a fully-resolved manifest.
 
     Capabilities are wholesale-replaced by the preset (atomic swap is the
-    whole point of presets) with one carve-out: ``library.paths`` from
-    init.json appends to the preset's library paths, deduped, preset
-    defaults first. Library paths are project-truth (where this agent's
-    extra skills live) — swapping presets should not erase them.
+    whole point of presets) with one carve-out: ``skills.paths`` from
+    init.json appends to the preset's skills paths, deduped, preset defaults
+    first. Skill paths are project-truth (where this agent's extra skills
+    live) — swapping presets should not erase them. Old ``library.paths`` is
+    accepted during the rename migration.
 
     Mutates ``data`` in place. No-op when ``manifest.preset.active`` is unset
     or when the manifest already has a literal ``llm`` and no preset block.
@@ -356,40 +356,44 @@ def materialize_active_preset(data: dict, working_dir: Path) -> None:
     # Capabilities are wholesale-replaced by the preset (atomic swap is the
     # whole point of presets). Two carve-outs preserve project-truth across
     # preset switches:
-    #   - library.paths: extras declared in init.json append to the preset's
-    #     defaults. Library paths are project-truth (where this agent's
-    #     extra skills live), not runtime-tier choices, so swapping presets
-    #     should not erase them.
+    #   - skills.paths: extras declared in init.json append to the preset's
+    #     defaults. Skill paths are project-truth (where this agent's extra
+    #     skills live), not runtime-tier choices, so swapping presets should
+    #     not erase them. Old library.paths is treated as skills.paths when
+    #     it appears during migration.
     # If you need to add another carve-out, do it here — keep the list short.
     init_caps = manifest.get("capabilities", {}) if isinstance(
         manifest.get("capabilities"), dict) else {}
-    init_lib_paths = (
-        init_caps.get("library", {}).get("paths", [])
-        if isinstance(init_caps.get("library"), dict) else []
-    )
+    init_skill_paths: list[str] = []
+    for cap_name in ("skills", "library"):
+        cfg = init_caps.get(cap_name)
+        if isinstance(cfg, dict) and isinstance(cfg.get("paths"), list):
+            for path in cfg.get("paths", []):
+                if isinstance(path, str) and path not in init_skill_paths:
+                    init_skill_paths.append(path)
 
     new_caps = preset_manifest.get("capabilities", init_caps)
-    if isinstance(new_caps, dict) and init_lib_paths:
-        # Make sure we have a library entry to merge into. If the preset
-        # didn't enable library, init.json's paths alone are enough to enable it.
-        lib_kwargs = new_caps.get("library", {})
-        if not isinstance(lib_kwargs, dict):
-            lib_kwargs = {}
-        preset_paths = lib_kwargs.get("paths", []) or []
+    if isinstance(new_caps, dict) and init_skill_paths:
+        # Make sure we have a skills entry to merge into. If the preset didn't
+        # enable skills, init.json's paths alone are enough to enable it.
+        skills_kwargs = new_caps.get("skills", {})
+        if not isinstance(skills_kwargs, dict):
+            skills_kwargs = {}
+        preset_paths = skills_kwargs.get("paths", []) or []
         # Order: preset paths first (the curated defaults), then init.json
         # extras unique to it. Dedupe by raw string — `~/foo` and the
         # absolute form count as different so the user's intent shows up
         # in info() output verbatim.
         merged: list[str] = list(preset_paths)
         seen = set(preset_paths)
-        for p in init_lib_paths:
+        for p in init_skill_paths:
             if p not in seen:
                 merged.append(p)
                 seen.add(p)
-        lib_kwargs = dict(lib_kwargs)
-        lib_kwargs["paths"] = merged
+        skills_kwargs = dict(skills_kwargs)
+        skills_kwargs["paths"] = merged
         new_caps = dict(new_caps)
-        new_caps["library"] = lib_kwargs
+        new_caps["skills"] = skills_kwargs
 
     manifest["capabilities"] = new_caps
     if preset_ctx is not None:
