+++ tests/test_check_caps.py
@@ -11,7 +11,7 @@ from lingtai.capabilities import get_all_providers
 def test_get_all_providers_returns_all_capabilities():
     result = get_all_providers()
     expected = {
-        "file", "bash", "web_search", "codex", "library",
+        "file", "bash", "web_search", "codex", "library", "skills",
         "vision", "avatar", "daemon",
     }
     assert expected == set(result.keys())
@@ -27,7 +27,7 @@ def test_get_all_providers_structure():
 
 def test_builtin_capabilities_have_empty_providers():
     result = get_all_providers()
-    builtins = ["file", "bash", "codex", "library", "avatar", "daemon"]
+    builtins = ["file", "bash", "codex", "library", "skills", "avatar", "daemon"]
     for name in builtins:
         assert result[name]["providers"] == [], f"{name} should have empty providers"
         assert result[name]["default"] == "builtin", f"{name} should default to builtin"
