+++ tests/test_library_knowledge.py
@@ -1,10 +1,8 @@
-"""Tests for codex capability — durable self-memory."""
+"""Tests for library capability — durable long-term knowledge."""
 from __future__ import annotations
 
 import json
 from unittest.mock import MagicMock
-
-import pytest
 
 from lingtai.agent import Agent
 
@@ -22,33 +20,67 @@ def make_mock_service():
 # ---------------------------------------------------------------------------
 
 
-def test_codex_setup_registers_tool(tmp_path):
+def test_library_setup_registers_tool(tmp_path):
     agent = Agent(
         service=make_mock_service(), agent_name="test", working_dir=tmp_path / "test",
-        capabilities=["codex"],
+        capabilities=["library"],
     )
-    assert "codex" in agent._tool_handlers
+    assert "library" in agent._tool_handlers
+    assert "codex" in agent._tool_handlers  # deprecated compatibility alias
     agent.stop(timeout=1.0)
 
 
-def test_codex_manager_accessible(tmp_path):
+def test_codex_capability_normalizes_to_library(tmp_path):
     agent = Agent(
         service=make_mock_service(), agent_name="test", working_dir=tmp_path / "test",
         capabilities=["codex"],
     )
-    mgr = agent.get_capability("codex")
+    try:
+        assert agent.get_capability("library") is not None
+        assert agent.get_capability("codex") is None
+        assert "library" in agent._tool_handlers
+        assert "codex" in agent._tool_handlers
+    finally:
+        agent.stop(timeout=1.0)
+
+
+def test_codex_tool_alias_uses_library_store(tmp_path):
+    agent = Agent(
+        service=make_mock_service(), agent_name="test", working_dir=tmp_path / "test",
+        capabilities=["library"],
+    )
+    try:
+        result = agent._tool_handlers["codex"]({
+            "action": "submit",
+            "title": "Compat",
+            "summary": "Old codex tool writes to renamed library.",
+        })
+        assert result["status"] == "ok"
+        prompt = agent._prompt_manager.read_section("library") or ""
+        assert "Compat" in prompt
+        assert (agent.working_dir / "codex" / "codex.json").is_file()
+    finally:
+        agent.stop(timeout=1.0)
+
+
+def test_library_manager_accessible(tmp_path):
+    agent = Agent(
+        service=make_mock_service(), agent_name="test", working_dir=tmp_path / "test",
+        capabilities=["library"],
+    )
+    mgr = agent.get_capability("library")
     assert mgr is not None
     agent.stop(timeout=1.0)
 
 
-def test_codex_independent_of_psyche(tmp_path):
-    """Codex is a separate capability; psyche is always-on as intrinsic."""
+def test_library_independent_of_psyche(tmp_path):
+    """Library is a separate capability; psyche is always-on as intrinsic."""
     agent = Agent(
         service=make_mock_service(), agent_name="test", working_dir=tmp_path / "test",
-        capabilities=["codex"],
+        capabilities=["library"],
     )
     assert "psyche" in agent._intrinsics
-    assert "codex" in agent._tool_handlers
+    assert "library" in agent._tool_handlers
     agent.stop(timeout=1.0)
 
 
@@ -60,9 +92,9 @@ def test_codex_independent_of_psyche(tmp_path):
 def test_submit_creates_entry(tmp_path):
     agent = Agent(
         service=make_mock_service(), agent_name="test", working_dir=tmp_path / "test",
-        capabilities=["codex"],
+        capabilities=["library"],
     )
-    mgr = agent.get_capability("codex")
+    mgr = agent.get_capability("library")
     result = mgr.handle({
         "action": "submit",
         "title": "TCP Retry Logic",
@@ -80,9 +112,9 @@ def test_submit_creates_entry(tmp_path):
 def test_submit_requires_title(tmp_path):
     agent = Agent(
         service=make_mock_service(), agent_name="test", working_dir=tmp_path / "test",
-        capabilities=["codex"],
+        capabilities=["library"],
     )
-    mgr = agent.get_capability("codex")
+    mgr = agent.get_capability("library")
     result = mgr.handle({"action": "submit", "summary": "s", "content": "c"})
     assert "error" in result
     agent.stop(timeout=1.0)
@@ -91,9 +123,9 @@ def test_submit_requires_title(tmp_path):
 def test_submit_enforces_limit(tmp_path):
     agent = Agent(
         service=make_mock_service(), agent_name="test", working_dir=tmp_path / "test",
-        capabilities={"codex": {"codex_limit": 2}},
+        capabilities={"library": {"library_limit": 2}},
     )
-    mgr = agent.get_capability("codex")
+    mgr = agent.get_capability("library")
     mgr.handle({"action": "submit", "title": "A", "summary": "s", "content": "c"})
     mgr.handle({"action": "submit", "title": "B", "summary": "s", "content": "c"})
     result = mgr.handle({"action": "submit", "title": "C", "summary": "s", "content": "c"})
@@ -111,9 +143,9 @@ def test_submit_without_content(tmp_path):
     """Title + summary alone is a valid entry — content is optional."""
     agent = Agent(
         service=make_mock_service(), agent_name="test", working_dir=tmp_path / "test",
-        capabilities=["codex"],
+        capabilities=["library"],
     )
-    mgr = agent.get_capability("codex")
+    mgr = agent.get_capability("library")
     result = mgr.handle({
         "action": "submit",
         "title": "A",
@@ -132,9 +164,9 @@ def test_submit_without_content(tmp_path):
 def test_view_returns_content(tmp_path):
     agent = Agent(
         service=make_mock_service(), agent_name="test", working_dir=tmp_path / "test",
-        capabilities=["codex"],
+        capabilities=["library"],
     )
-    mgr = agent.get_capability("codex")
+    mgr = agent.get_capability("library")
     r = mgr.handle({"action": "submit", "title": "X", "summary": "s", "content": "full content here"})
     result = mgr.handle({"action": "view", "ids": [r["id"]]})
     assert result["status"] == "ok"
@@ -147,9 +179,9 @@ def test_view_returns_content(tmp_path):
 def test_view_with_include_supplementary(tmp_path):
     agent = Agent(
         service=make_mock_service(), agent_name="test", working_dir=tmp_path / "test",
-        capabilities=["codex"],
+        capabilities=["library"],
     )
-    mgr = agent.get_capability("codex")
+    mgr = agent.get_capability("library")
     r = mgr.handle({
         "action": "submit", "title": "X", "summary": "s",
         "content": "main", "supplementary": "extra material",
@@ -165,9 +197,9 @@ def test_view_with_include_supplementary(tmp_path):
 def test_view_invalid_id(tmp_path):
     agent = Agent(
         service=make_mock_service(), agent_name="test", working_dir=tmp_path / "test",
-        capabilities=["codex"],
+        capabilities=["library"],
     )
-    mgr = agent.get_capability("codex")
+    mgr = agent.get_capability("library")
     result = mgr.handle({"action": "view", "ids": ["nope"]})
     assert "error" in result
     agent.stop(timeout=1.0)
@@ -177,9 +209,9 @@ def test_filter_and_export_actions_rejected(tmp_path):
     """Removed actions return error, not silent no-op."""
     agent = Agent(
         service=make_mock_service(), agent_name="test", working_dir=tmp_path / "test",
-        capabilities=["codex"],
+        capabilities=["library"],
     )
-    mgr = agent.get_capability("codex")
+    mgr = agent.get_capability("library")
     for action in ("filter", "export"):
         result = mgr.handle({"action": action})
         assert "error" in result, f"{action} should be rejected"
@@ -195,9 +227,9 @@ def test_filter_and_export_actions_rejected(tmp_path):
 def test_consolidate(tmp_path):
     agent = Agent(
         service=make_mock_service(), agent_name="test", working_dir=tmp_path / "test",
-        capabilities=["codex"],
+        capabilities=["library"],
     )
-    mgr = agent.get_capability("codex")
+    mgr = agent.get_capability("library")
     r1 = mgr.handle({"action": "submit", "title": "A", "summary": "s1.", "content": "c1"})
     r2 = mgr.handle({"action": "submit", "title": "B", "summary": "s2.", "content": "c2"})
     result = mgr.handle({
@@ -223,9 +255,9 @@ def test_consolidate(tmp_path):
 def test_delete(tmp_path):
     agent = Agent(
         service=make_mock_service(), agent_name="test", working_dir=tmp_path / "test",
-        capabilities=["codex"],
+        capabilities=["library"],
     )
-    mgr = agent.get_capability("codex")
+    mgr = agent.get_capability("library")
     r1 = mgr.handle({"action": "submit", "title": "A", "summary": "s.", "content": "c"})
     r2 = mgr.handle({"action": "submit", "title": "B", "summary": "s.", "content": "c"})
     result = mgr.handle({"action": "delete", "ids": [r1["id"]]})
@@ -243,7 +275,7 @@ def test_delete(tmp_path):
 
 
 def test_schema_has_all_fields():
-    from lingtai.core.codex import get_schema
+    from lingtai.core.library import get_schema
     SCHEMA = get_schema("en")
     actions = SCHEMA["properties"]["action"]["enum"]
     assert set(actions) == {"submit", "view", "consolidate", "delete"}
@@ -266,15 +298,15 @@ def test_schema_has_all_fields():
 
 
 def test_id_deterministic():
-    from lingtai.core.codex import CodexManager
-    id1 = CodexManager._make_id("hello", "2026-03-16T00:00:00Z")
-    id2 = CodexManager._make_id("hello", "2026-03-16T00:00:00Z")
+    from lingtai.core.library import LibraryManager
+    id1 = LibraryManager._make_id("hello", "2026-03-16T00:00:00Z")
+    id2 = LibraryManager._make_id("hello", "2026-03-16T00:00:00Z")
     assert id1 == id2
     assert len(id1) == 8
 
 
 def test_id_differs_by_content():
-    from lingtai.core.codex import CodexManager
-    id1 = CodexManager._make_id("hello", "2026-03-16T00:00:00Z")
-    id2 = CodexManager._make_id("world", "2026-03-16T00:00:00Z")
+    from lingtai.core.library import LibraryManager
+    id1 = LibraryManager._make_id("hello", "2026-03-16T00:00:00Z")
+    id2 = LibraryManager._make_id("world", "2026-03-16T00:00:00Z")
     assert id1 != id2
