+++ tests/test_prompt.py
@@ -20,7 +20,7 @@ def test_build_system_prompt_with_sections():
 def test_rules_renders_after_covenant_and_tools():
     """Section order is grouped by mutation frequency for cache stability:
     Batch 1 (immovable, prefix-cacheable) — principle, covenant, tools, substrate, ...
-    Batch 2 (rarely mutated)              — rules, brief, library, codex, ...
+    Batch 2 (rarely mutated)              — rules, brief, skills, library, ...
 
     So both ``covenant`` and ``tools`` precede ``rules`` in the rendered
     prompt, since adjusting rules at runtime should invalidate as little
